% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_stats.R
\name{get_player_stats}
\alias{get_player_stats}
\title{Get Player Stats}
\usage{
get_player_stats()
}
\value{
DPLYR tbl_df
}
\description{
Get Player Stats
}
\examples{
player_stats <- get_player_stats()
}
