% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_mini.R
\name{get_results_mini}
\alias{get_results_mini}
\title{Get Results Mini}
\usage{
get_results_mini()
}
\value{
DPLYR tbl_df
}
\description{
Get Results Mini
}
\examples{
results_mini <- get_results_mini()
}
