% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subs_and_reds.R
\name{get_subs_and_reds}
\alias{get_subs_and_reds}
\title{Get subs and reds}
\usage{
get_subs_and_reds()
}
\value{
DPLYR tbl_df
}
\description{
Get subs and reds
}
\examples{
subs_and_reds <- get_subs_and_reds()
}
